fix(client): added custom validation to the same file
