+++ packages/amplication-client/src/Entity/EntityForm.tsx
@@ -10,7 +10,8 @@ import NameField from "../Components/NameField";
 import { Form } from "../Components/Form";
 import FormikAutoSave from "../util/formikAutoSave";
 import { USER_ENTITY } from "./constants";
-import { validateSchemaAndEqualEntities } from "../util/customValidations";
+import { validate } from "../util/formikValidateJsonSchema";
+import { isEqual } from "../util/customValidations";
 
 export type EntityInput = Omit<models.Entity, "fields" | "versionNumber">;
 
@@ -32,6 +33,27 @@ const NON_INPUT_GRAPHQL_PROPERTIES = [
   "__typename",
 ];
 
+const FORM_SCHEMA = {
+  required: ["name", "displayName", "pluralDisplayName"],
+  properties: {
+    displayName: {
+      type: "string",
+      minLength: 2,
+    },
+    name: {
+      type: "string",
+      minLength: 2,
+    },
+    pluralDisplayName: {
+      type: "string",
+      minLength: 2,
+    },
+  },
+};
+
+const EQUAL_PLURAL_DISPLAY_NAME_AND_NAME_TEXT =
+  "Name and plural display names cannot be equal. The ‘plural display name’ field must be in a plural form and ‘name’ field must be in a singular form";
+
 const CLASS_NAME = "entity-form";
 
 const EntityForm = React.memo(({ entity, applicationId, onSubmit }: Props) => {
@@ -49,7 +71,14 @@ const EntityForm = React.memo(({ entity, applicationId, onSubmit }: Props) => {
     <div className={CLASS_NAME}>
       <Formik
         initialValues={initialValues}
-        validate={validateSchemaAndEqualEntities}
+        validate={(values) => {
+          if (isEqual(values.name, values.pluralDisplayName)) {
+            return {
+              pluralDisplayName: EQUAL_PLURAL_DISPLAY_NAME_AND_NAME_TEXT,
+            };
+          }
+          return validate(values, FORM_SCHEMA);
+        }}
         enableReinitialize
         onSubmit={onSubmit}
       >
